tddev/requirements.txt
# TDDev - Test-Driven Development Framework for Web Application Generation
# Based on arXiv:2509.25297v2

# LLM API Clients
openai>=1.0.0
anthropic>=0.18.0

# Browser Automation & Testing
browser-use>=0.1.0
playwright>=1.40.0
selenium>=4.0.0

# Web Framework Support
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# File & Code Processing
GitPython>=3.1.0
jinja2>=3.1.0
pyyaml>=6.0.0
toml>=0.10.0

# Image Processing (Multimodal Support)
pillow>=10.0.0

# Process Management
psutil>=5.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Logging & Monitoring
loguru>=0.7.0
tqdm>=4.65.0

# Testing & Validation
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Local LLM Support
# transformers>=4.30.0
# torch>=2.0.0
